# StrnadiAPI-openapi.yaml
openapi: 3.0.1
info:
  title: Strnadi API
  description: |
    HTTP API used by the Strnadi applications to manage user accounts, bird recordings,
    articles and other shared resources. The specification documents the behavior of the
    existing ASP.NET controllers so it can be imported into Swagger UI or other OpenAPI tooling.
  version: 1.0.0
servers:
  - url: https://api.strnadi.cz
    description: PROD API
  - url: https://devapi.strnadi.cz
    description: DEV API
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      description: Returns basic information for every registered user. Only administrators may access this endpoint.
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserModel'
        '400':
          description: The request did not contain a bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The caller is not an administrator or supplied an invalid JWT.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load users from the repository.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/get-id:
    get:
      tags: [Users]
      summary: Get the identifier of the current user
      description: Extracts the user id from the supplied JWT.
      responses:
        '200':
          description: Identifier of the authenticated user.
          content:
            application/json:
              schema:
                type: integer
                minimum: 1
        '400':
          description: Missing bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The supplied JWT was invalid or the user no longer exists.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get profile by identifier
      description: |
        Returns a user profile. When the caller is unauthenticated only non-sensitive fields are
        returned (the email is omitted). Authenticated callers receive their full profile and
        administrators may view the email of any account.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      security: []
      responses:
        '200':
          description: User profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
        '401':
          description: The supplied JWT was invalid for viewing protected fields.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No user with the supplied identifier exists.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Users]
      summary: Update profile information
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserModel'
      responses:
        '200':
          description: The profile was updated.
        '400':
          description: The request did not contain a bearer token or referenced a different user.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The supplied JWT is invalid.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The profile could not be updated.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Users]
      summary: Delete a user account
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: The account has been deleted.
        '400':
          description: The request did not contain a bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The caller is not allowed to delete this user.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The user could not be deleted because it does not exist.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/verify-email:
    get:
      tags: [Users]
      summary: Finalize email verification
      description: Confirms a verification link and redirects to the configured front-end application.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: jwt
          in: query
          required: true
          description: Verification JWT included in the email message.
          schema:
            $ref: '#/components/schemas/JwtToken'
      security: []
      responses:
        '301':
          description: Redirect to the front-end confirmation page.
          headers:
            Location:
              description: Target redirection URL.
              schema:
                type: string
                format: uri
        '401':
          description: The verification token is invalid.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/change-password:
    patch:
      tags: [Users]
      summary: Change the password for the current user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password updated.
        '400':
          description: Missing bearer token or attempting to change another user's password.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid JWT supplied.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: The password could not be changed due to a server error.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/exists:
    get:
      tags: [Users]
      summary: Check whether a user exists
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
      security: []
      responses:
        '200':
          description: The user does not exist and the identifier may be used.
        '400':
          description: Neither `email` nor `userId` was provided.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A user already exists with the supplied identifier.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/upload-profile-photo:
    post:
      tags: [Users]
      summary: Upload or replace a user profile photo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfilePhotoModel'
      responses:
        '200':
          description: Photo uploaded successfully.
        '400':
          description: Missing bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid JWT supplied.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The photo could not be saved.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/get-profile-photo:
    get:
      tags: [Users]
      summary: Download the stored profile photo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      security: []
      responses:
        '200':
          description: Base64 encoded profile photo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePhotoModel'
        '404':
          description: The user does not have a stored profile photo.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: string
      description: Human readable message describing the error.
    JwtToken:
      type: string
      description: Encoded JSON Web Token.
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    UpdateUserModel:
      type: object
      properties:
        nickname:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        postCode:
          type: integer
          nullable: true
    ChangePasswordRequest:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          format: password
    UserProfilePhotoModel:
      type: object
      required: [format, photoBase64]
      properties:
        format:
          type: string
          description: File extension (for example `jpg`).
        photoBase64:
          type: string
          format: byte
          description: Base64 encoded profile picture.
...
    UserModel:
      type: object
      required: [id, email, firstName, lastName, creationDate]
      properties:
        id:
          type: integer
          minimum: 1
        nickname:
          type: string
          nullable: true
        email:
          type: string
          format: email
        password:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        postCode:
          type: integer
          nullable: true
        city:
          type: string
          nullable: true
        creationDate:
          type: string
          format: date-time
        isEmailVerified:
          type: boolean
          nullable: true
        consent:
          type: boolean
          nullable: true
        role:
          type: string
          nullable: true